web3-toolkit/pyproject.toml
[tool.poetry]
name = "web3-python-toolkit"
version = "0.1.0"
description="toolkit for web3"
authors=["steinkirch"]
readme = "README.md"
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = "3.11.1"
ethereum-etl = "2.1.2"
python-dotenv = "0.21.0"
web3 = "5.31.3"
pandas = "^1.5.2"

[tool.poetry.group.dev.dependencies]
tox = "^4.0.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
